Move get{Data,Markup}FromTree into new @apollo/react-ssr package.
The RenderPromises class needs to remain in packages/hooks/src/ssr because
it relies on the private QueryData type, but this reoganization means we
no longer call require("react-dom/server") anywhere except in the new
@apollo/react-ssr package.

Should address https://github.com/apollographql/apollo-feature-requests/issues/118.
