+++ packages/components/package.json
@@ -59,6 +59,7 @@
     "access": "public"
   },
   "devDependencies": {
+    "@apollo/react-ssr": "file:../ssr",
     "jest": "^24.8.0",
     "rollup": "1.16.2",
     "tsc-watch": "^2.2.1",
