+++ packages/components/src/__tests__/ssr/getDataFromTree.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/server';
 import ApolloClient from 'apollo-client';
 import { ApolloProvider, getApolloContext } from '@apollo/react-common';
-import { getDataFromTree } from '@apollo/react-hooks';
+import { getDataFromTree } from '@apollo/react-ssr';
 import gql from 'graphql-tag';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
 import { mockSingleLink } from '@apollo/react-testing';
