+++ packages/components/src/__tests__/ssr/server.test.tsx
@@ -11,7 +11,7 @@ import {
   GraphQLID
 } from 'graphql';
 import { ApolloProvider } from '@apollo/react-common';
-import { renderToStringWithData } from '@apollo/react-hooks';
+import { renderToStringWithData } from '@apollo/react-ssr';
 import gql from 'graphql-tag';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
 import { Query } from '@apollo/react-components';
