+++ packages/components/src/index.ts
@@ -9,10 +9,4 @@ export { Query } from './Query';
 export { Mutation } from './Mutation';
 export { Subscription } from './Subscription';
 
-export {
-  getMarkupFromTree,
-  getDataFromTree,
-  renderToStringWithData
-} from '@apollo/react-hooks';
-
 export * from './types';
