+++ packages/ssr/config/rollup.config.js
@@ -1,0 +1,5 @@
+import { rollup } from '../../../config/rollup.config';
+
+export default rollup({
+  name: 'ssr'
+});
