+++ packages/ssr/config/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../../config/tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "../lib"
+  },
+  "include": ["../src/**/*"],
+  "exclude": ["../src/**/__tests__/**/*", "../src/**/__mocks__/**/*"]
+}
